requirements.txt
vantage6-algorithm-tools==4.9.1
matplotlib==3.10.1
pandas
pydantic